# web/lib/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AapatMitra API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/webhooks/sms/{path_secret}": {
   "post": {
    "tags": [
     "comms"
    ],
    "summary": "Webhook Sms",
    "operationId": "webhook_sms_api_v1_webhooks_sms__path_secret__post",
    "parameters": [
     {
      "name": "path_secret",
      "in": "path",
      "required": true,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Path Secret"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/webhooks/sms": {
   "post": {
    "tags": [
     "comms"
    ],
    "summary": "Webhook Sms",
    "operationId": "webhook_sms_api_v1_webhooks_sms_post",
    "parameters": [
     {
      "name": "path_secret",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Path Secret"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/notifications/{notification_id}/opened": {
   "post": {
    "tags": [
     "comms"
    ],
    "summary": "Notification Opened",
    "operationId": "notification_opened_api_v1_notifications__notification_id__opened_post",
    "parameters": [
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Notification Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/notifications": {
   "get": {
    "tags": [
     "comms"
    ],
    "summary": "Notifications",
    "operationId": "notifications_api_v1_notifications_get",
    "parameters": [
     {
      "name": "unread",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Unread"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Notifications Api V1 Notifications Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}